[BUGFIX] Use ADMCMD_prev to keep live version in workspace split view
Previously, when using the "Split Preview" for comparing a workspace
version, and a live workspace, in iframes, the workspace version
received the GET Parameter "&ADMCMD_prev=1", and the live workspace
the "ADMCMD_noBeUser=1" (= no backend user is logged in) for THESE
requests.

Navigating through the live workspace, the ADMCMD_* GET parameters are
lost in both variants.

This change replaces the ADMCMD_noBeUser=1 with ADMCMD_prev=LIVE
making the options consistent naming-wise.

In addition, when ADMCMD_prev=LIVE is set, the current BE_USER
is kept but the workspace is set to "0" temporary (for this request),
allowing for e.g. the admin panel to continue to work.

When ADMCMD_prev=LIVE or ADMCMD_prev=IGNORE is set, the parameter
is added to config.linkVars, allowing to keep using the ADMCMD_prev
when browsing through the workspace and live versions in the iframes.

Resolves: #34199
Releases: master, 10.4
Change-Id: Ifa732d56af1af1d6f01f00b6b4c035b6fdbc02ce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65198
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>