[BUGFIX] Correctly build query string without cHash
For building URIs using the Symfony routing,
array URL arguments are deflated (foo[bar] -> foo__bar).

In case the final URI contains a remaining query string,
the arguments get inflated again. However the
query string of the URI is currently only updated, when
at least one of the remaining arguments require a cHash.

In case remaining arguments exist, the query string
needs to be always updated, whether the URI requires
a cHash or not.

Resolves: #92183
Releases: master, 10.4, 9.5
Change-Id: I1966b4b1e8e09716733b3357922cf857543c14e4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65596
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Helmut Hummel <typo3@helhum.io>