[TASK] Remove no wrap divider in imageorient
The divider element in the TCA option for "tt_content.imageorient"
does not reflect the current state of the functionality
anymore and can be removed, since the "no wrap" is not
very appealing for editors to understand, and other options
(e.g. above) are also without wrapping, so this div
element is removed.

Resolves: #50570
Releases: master
Change-Id: If41b3baf2ca768521dcabc08b2e9287b9a792fa9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65579
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>