[!!!][FEATURE] Always initialize frontend groups after FE user
In previous TYPO3 versions (due to historic reasons) the frontend
groups were always resolved within TSFE when a page and the
rootline was resolved.

However, this left the actual Frontend User, which is initialized
at the very beginning of a frontend request, in an incomplete state:
A user was (correctly) found and "logged in", but the groups were
resolved at a later point.

This was due to the fact that the Admin Panel allowed to
"include hidden records" which also considered fe_groups,
and thus be set later-on.

This change now moved the resolving of the groups (and setting
the right frontend.user aspect) right after the user resolving.

This means that the groups are now available much earlier, and not
bound to the TSFE instance anymore, allowing to use Middlewares much
more professionally without depending on TSFE for custom Routing / APIs.

Future options:
It would even be possible to filter out PageRouter pages that are
not available, which would make the Router itself faster.

Resolves: #92562
Releases: master
Change-Id: Ia522697433049b0e549f3c65caf6757053ff37e1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66066
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>