[!!!][TASK] Do not handle deleted records in reference index
Keeping relation information for soft deleted records
in sys_refindex is quite useless: When records are
undeleted, the reference index is updated to recreate
rows. Additionally, the deleted handling was incomplete,
especially if a record on the 'right' side of the relation
got deleted.

The patch drops the deleted field from the table and
prevents adding relations for soft deleted rows.
Furthermore, visibility of a couple of properties and
methods is changed to protected and a series of not used
methods is removed.

Note the hash sums change due to the removal of the deleted
field. The CLI command "bin/typo3 referenceindex:update"
will update hashes and will drop obsolete deleted rows.

Change-Id: I58d7a904a6b4c555529b7c70e45d56ccb498f77f
Resolves: #93029
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66968
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>