[!!!][TASK] Rework user group fetching for frontend users
This change adapts the frontend user resolving to use
the same functionality as in backend users, which wasn't
the case in the past 15 years.

The newly introduced GroupResolver now also resolves
groups for frontend users recursively, making the "getGroupsFE"
and "authGroupsFE" functionality obsolete.

The AuthenticationService->getGroups methods are removed,
also the properties userTS, userData_change and TSdataArray have
been made protected, as it is only used for internal purposes.

The "userTSUpdated" flag has been removed and the property $TSdataArray
has been marked as internal.

Resolves: #93108
Releases: master
Change-Id: I630be39e855d45802c0236a0a38acbc07e0dd812
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67177
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>