[TASK] Streamline backend UriBuilder dependency injection
The Router dependency should not be optional and rather always be
injected, and in reality it always was, as the nullable fallback
to GeneralUtility::makeInstance was only used in unit tests.

The UriBuilder factory is now provided via the backend
service provider. That is because it became an ondemand
dependency for the backend request handler in #93674.
We now properly inject it via constructor which makes it
a lowlevel service (a service that needs to defined via
service provider to be available when TYPO3 has not yet
been installed).

Releases: master
Resolves: #94083
Change-Id: I75b39cdae8c41d1ce38860c50713d749bbe8e348
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69056
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Benni Mack <benni@typo3.org>