[TASK] Explicitly set ContentObjectRenderer in USER cObj
Handing around ContentObjectRenderer in the frontend
is tricky and intransparent. Especially class instances
created via ContentObjectRenderer->callUserFunction()
get an instance of it set through public property $cObj
in the consuming class. This is typically the case for
USER and USER_INT type frontend plugins.

The patch deprecates using this public property and
calls the method setContentObjectRenderer() instead.

This explicit method call is much easier to see,
break point and follow. The state handling of
ContentObjectRenderer is far from obvious, the
change helps debugging and refactoring this further.

Change-Id: Ib6ca11e7d6660dae0e82ad08ad1f9b39095e99da
Resolves: #94956
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70717
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>