[BUGFIX] Keep original request language for shortcuts
In #89871 (65dc060) a fix was introduced, that
forces the link generation to use the requested
language of shortcuts instead of the current
language of the frontend.

In addition, another call to getPageOverlay was
added. This is in fact not necessary, as overlaying
is already done implicitly in getPageShortcut. This
introduced a falsy behaviour, where shortcuts
without the current language available would create
a URL with the entry point of the fallback language.

This can be fixed by simply using the request
language from the original $menuItem variable before
processing.

Resolves: #94677
Related: #89871
Releases: master, 10.4
Change-Id: I5c9bd0439d821d5516a5cb58cdf53df6b63c3aab
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70836
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>