[TASK] Cover FE index.php id/type parameter requests with tests
Core has been working with id and type as query
parameter to display FE pages and switch between
defined TypoScript page definitions. Speaking url
has been integrated for v9 through the new routing
feature to do it in a more determistic way as in
comparision to ext:realurl.

As there might be instances using id/type query
parameters for ajax call switches (and eventelly
for eid scrips) in the frontend, normalization has
been implemented to keep this option alive (#88028).

Recently there has been a try to fix a bug for so
some use cases, which broke this support (#94537).

For future it should be considered if it is possible
to drop this option in the future to free id/type
query arguments for instances in the frontend.

There a several issues in this area, for example
broken subfolder installation support and id/type
query parameter on a full speaking url overrule
it with selected id, which is not expected.

To validate that the test fails if support is
broken change in PageRouter following line (150):

$urlPath = str_replace($scriptName, '', $urlPath);

to

$urlPath = str_replace( '/' . $scriptName,
    '', $urlPath);

Build/Scripts/runTests.sh -s functional \
typo3/sysext/frontend/Tests/Functional/SiteHandling/\
SiteRequestTest.php

This patch acts as pre-patche to fix these.

Resolves: #95096
Related: #95096
Releases: master, 10.4
Change-Id: Iff97f04d7bf7e7b2462425d830e0451a8d7a8686
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71189
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>