[TASK] Avoid referencing public assets across packages
Since TYPO3 core packages (extensions) are independently
published and might be installed in arbitrary file locations
in the near future, referencing assets in package A
from a CSS file package B will likely not work.

Therefore all assets that are referenced from a CSS
file are now moved to the package that contains this CSS file.

Also references of assets in CSS that already were located
in the same package are cleaned up to not assume a hard coded
path name (breaking out of Resources/Public directory).

Resolves: #95372
Releases: master
Change-Id: I6d7d7557d8ea7b293c81b50bc4172019a088f7b7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71332
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benni Mack <benni@typo3.org>