[BUGFIX] Resolve shortcuts in typolink directly
Currently, HMENU generates the direct destination for
a shortcut URL, but typolink does not. This is an
inconsistency, and since HMENU is actually using typolink
now as well (in 4.x this was different), the resolving
of shortcuts can now be moved into the typolink/link building
functionality.

Page Shortcuts are now resolved (recursively!) at a single point for
all links, except for RANDOM_SUBPAGE variants, as this is sorted
out during runtime when hitting the RANDOM_SUBPAGE shortcut page.

Resolves: #80113
Resolves: #95947
Releases: main, 11.5
Change-Id: Ib6204ff1b0ecb699ec87f5154dc18c974676e65c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72772
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>