[BUGFIX] Set fe_user email field to eval "email"
Currently the fe_user field email has a eval value
of "trim". So it does not verify the given email.

This has been fixed by setting eval to "email" and
trim the email before validation. On top of that
the validation flash message now shows the correct
value that was entered.

Resolves: #95459
Releases: master
Change-Id: I274ae040da9fe8d03a56e5ba9b4f17a828da57f7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71483
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>