[BUGFIX] Prevent htmlSanitize deprecation in ContentObjectRendererTest
htmlSanatize configuration has been marked as mandatory for v12 with
corresponding deprecation message with #94786 if htmlSanatize is not
set at all.

This triggers deprecation messages in ContentObjetRendererTests which
should not trigger depcrecation messages as tests should test for
correct rendering.

This patch sets htmlSanatize option for this tests to prevent these
messages.

Resolves: #95611
Related: #94786
Releases: master
Change-Id: I3222e5df98cdc2e56a86cefd7328e2b51a352f80
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71602
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>