[TASK] Simplify checkbox items TCA
TCA shares same keys across different types,
and so does "items" with type check, radio
and select. The value is usually set for the
array key "1". But type check is special in
that regard: the value is stored as a bitset.
This means, the value is completely unused
and shall not be provided for checkboxes.

1. All occurrences of "items => ['1' => '']"
   removed.

2. When checkboxes only have 1 item and no
   configuration whatsoever, "items" can be
   omitted. These constructs are gone now:

   'items' => [
       [
            0 => '',
            1 => '',
       ],
    ]

Resolves: #95876
Releases: master
Change-Id: I3deb96498e3afb8823cd6397e57e268c68774276
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72056
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>