[BUGFIX] Do not overwrite previously set preview mode
The frontend preview aspect may be set from the admin panel
or the frontend extension or even from a third-party extension
in a middleware. The previous implementation overwrote the preview
settings on every initialisation, resulting in lost preview
information. Instead of overwriting the setting, this change takes
the current preview value into account - if the preview was enabled
by a previous instance this is kept.

Additionally, the logic to set the preview flag in both frontend
and admin panel has been streamlined.

Resolves: #91225
Releases: main, 11.5
Change-Id: I9f642bca1928638cac5eb5c1468710f5fad28137
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72872
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>