[BUGFIX] Use typolink consistently for makelinks functionality
TYPO3's "makelinks" functionality parses for in-text
external URLs and mailto links. This change uses
Typolink under the hood now to consistently check
for invalid content, and have the chance to use
all hooks in typolink for such links as well.

This way, typolink is used, but makelinks is
executed AFTER possible resolved <a> tags
avoiding double-calling parseFunc when <a> tag
processing is configured.

Resolves: #93012
Releases: main, 11.5
Change-Id: I574915e715793f6eccad1bdfc216e82e329179dd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72871
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>