[!!!][TASK] Remove HMENU.collapse option
When the "collapse" option in the HMENU ContentObject
is activated, any active page links to its actual parent
page. This option was never touched since TYPO3 4.0
and stems from the time when TYPO3 still included JSMENU
and GMENU_LAYERS functionalities, which is not needed anymore
as more modern solutions such as MenuProcessor for Fluid
are more-widespread.

Resolves: #96517
Releases: main
Change-Id: I7f16369efaaeb17cc9d510307df808aefedd5945
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72939
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>