[BUGFIX] Avoid calling trim(null) in ContentObjectRenderer
'ContentObjectRenderer->stdWrap()' calls subfunction starting
with 'stdWrap_' which may not only return strings, but also int,
boolean, null and so on, thus eventually feeding 'trim()' with
null or a non string value to which emits a PHP8.0 warning and
an type error on PHP8.1.

To mitigate this string cast is added on the value for the
'trim(...)' call to be close to the spot and do not change
behaviour for now.

Releases: main, 11.5
Resolves: #96564
Change-Id: I347b59065106285860709f69400a1fc89bce2517
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73052
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>