[BUGFIX] Use Event.currentTarget in default_frontend.js
Event.target identifies the element on which the event
occurred and may a descendant of the element to which
the event handler has been attached.

Event.currentTarget instead always refers to the
element to which the event handler has been attached
and is therefore now used as source for the datasets.

Resolves: #96668
Releases: main, 11.5
Change-Id: Ic0231194a488feba8aaf9379d95b07b80d6a7e02
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73170
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>