[TASK] Remove superfluous function call
The function call seems superfluous, the condition tests if
the array key exists, so the first parameter should be
empty, the called function cObjGet tests if the first
parameter is an array, if not, an empty string is returned.
So i think we could use the empty string directly.

Releases: main, 11.5
Resolves: #96805
Change-Id: Ib0fcf38fd67ed66cc7078530a215629a2c2cde1f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73788
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>