[TASK] Use dependency injection in PageRenderer
Add dependency injection to PageRenderer and it's direct
dependencies. Most prominently this avoids the static
assets-cache injection from TYPO3 bootstrap and refactors to
regular injection instead.
ResourceCompressor is adapted to be delay it's initialization
to be injectable without immediate side-effects.

Backend Router and UriBuilder are not injected because they are
a) only needed in backend context and
b) should ideally be moved out of the PageRenderer
   into EXT:backend anyway.

Resolves: #97030
Releases: main
Change-Id: I8e72e5f87a095372439c3cd8106318ae272c6ce0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73594
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>