[TASK] Streamline AbstractUserAuthentication unpack_uc()
The "uc" value of a FE User / BE User is now
moved in the right spot so it does not need
to be called explicitly anymore.

"uc" is now unpacked right before MFA is needed during
the login process, so MFA can access the "uc".

As "uc" is needed after group resolving, it is added
at the right place for all code places, making
it obsolete to be called in user-land code.

In addition the "uc" property "firstLoginTimeStamp"
is removed which was used for TYPO3 v4.3 to show
a donation banner, which was removed again years ago.

Resolves: #97179
Releases: main
Change-Id: I6f44e9496174cb9a3da73c79ab96ffc27059daf6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73907
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Daniel Gorges <daniel.gorges@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Daniel Gorges <daniel.gorges@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>