[BUGFIX] Ensure imageLinkWrap works with UTF-8 filenames
The unset($info['originalFile']) introduced in #90565
broke an access to that array key below in the method.
This leads to the situation that files with UTF-8
characters are not found anymore since the imageLinkWrap()
method is fed with an encoded version of the filename
and not the original file object anymore.
The patch parks $info['originalFile'] in a variable
to access it for imageLinkWrap() again.

Releases: main, 11.5
Resolves: #93716
Resolves: #90757
Related: #90565
Change-Id: I2471283bd6825c093745b2331a12737de00c105f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73983
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>