[!!!][TASK] Remove eval=int from type=datetime
TCA type "datetime" allowed the value to be a native
type (with `dbType`), a int with `eval=int` or a
simple string. Since the value is always a timestamp
- in case no native type is used - the `eval=int`
option is removed and the value is always handled
as integer.

Note: Due to various problems with migration of existing
fields, this patch does not yet enforce the database
field to be of type "integer". This will be done step
by step in further patches, e.g.
https://review.typo3.org/c/Packages/TYPO3.CMS/+/74281

Resolves: #97358
Releases: main
Change-Id: Ia8297854fb75241660ffa2a9df850b22dd7a0b84
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74205
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>